Override the docker image we are using to build the pipeline (#81)
Why

We cannot publish a docker container for this pipeline due to the needed docker image being unavailable. Upgrading ingest-utils to 2.1.8 causes codegen issues on our usage of the scala reserved keyword trait for one of the entities we model. Until we can figure that out, let's try just overriding the docker image in the build.sbt settings.

This PR

Overrides the dockerBaseImage setting to point at the CE graalvm image.