[DSPDC-1858] Workaround jade API change (#89)
The semantics of the total field has changed in the Jade API enumerateSnapshots response. Before it reflected the total after filtering, it now reflects the total available. There is a new field, filteredTotal that we should begin using.

Since the currently published version of the Jade API client does not support this new field and the clinvar ingest is blocked until we get this resolved, I have switched to making a raw HTTP request via the python reequests library.

I have tested this locally using the prod argo config and credentials, no errors and the result is what we expect.