[DSPDC-1861] Put onExit hook inside of template, add retries (#92)
Why

The clinvar ingest fails to notify when a snapshot is cut, and also fails out when Jade gives back a 500 during the snapshotting process.

This PR

Moves the slack notification for a new snapshot into the template itself, instead of the callsite.
Adds argo retries around the Jade check latest snapshot and submit snapshot calls. I'd like to add a retry to the poll snapshot piece as well, but that is brought in via monster-helm and will need to be a PR in that repo.